+++ tests/Fixtures/Middlewares/ClassMiddleware.php
@@ -10,8 +10,10 @@ use Karewan\KnRoute\IMiddleware;
 #[Attribute(Attribute::TARGET_CLASS)]
 class ClassMiddleware implements IMiddleware
 {
-	public function handle(): void
+	public function before(): void
 	{
 		echo 'class>';
 	}
+
+	public function after(): void {}
 }