+++ tests/Fixtures/Middlewares/GlobalMiddleware.php
@@ -10,9 +10,11 @@ class GlobalMiddleware implements IMiddleware
 {
 	public static string $executionOrder = '';
 
-	public function handle(): void
+	public function before(): void
 	{
 		self::$executionOrder .= 'first';
 		\Karewan\KnRoute\header('X-Global-Middleware: true');
 	}
+
+	public function after(): void {}
 }