+++ tests/Fixtures/Middlewares/SecondGlobalMiddleware.php
@@ -8,9 +8,11 @@ use Karewan\KnRoute\IMiddleware;
 
 class SecondGlobalMiddleware implements IMiddleware
 {
-	public function handle(): void
+	public function before(): void
 	{
 		GlobalMiddleware::$executionOrder .= ',second';
 		\Karewan\KnRoute\header('X-Global-Order: ' . GlobalMiddleware::$executionOrder);
 	}
+
+	public function after(): void {}
 }